+++ pik/core/shortcuts/__init__.py
@@ -1,6 +1,12 @@
-from .model_objects import get_object_or_none, validate_and_create_object, \
-    validate_and_update_object, update_or_create_object
+from .model_objects import (
+    get_object_or_none, validate_and_create_object,
+    validate_and_update_object, update_or_create_object)
+from .request import get_current_request
 
 __all__ = [
-    'get_object_or_none', 'validate_and_create_object',
-    'validate_and_update_object', 'update_or_create_object']
+    'get_object_or_none',
+    'validate_and_create_object',
+    'validate_and_update_object',
+    'update_or_create_object',
+    'get_current_request',
+]